+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- S3 sources: add support for custom regions. ([#448](https://github.com/getsentry/symbolicator/pull/448))
 - Add support for BCSymbolMap auxiliary files which when present on the symbol server will automatically resolve obfuscated symbol names ([#403](https://github.com/getsentry/symbolicator/pull/403))
 - S3 sources: add support for using AWS container credentials provided by IAM task roles. ([#417](https://github.com/getsentry/symbolicator/pull/417))
 - Implement new Symbolication Quality Metrics. ([#426](https://github.com/getsentry/symbolicator/pull/426))